+++ src/pages/Gallery.js
@@ -1,5 +1,5 @@
 import React from "react";
-import ImageList from '../components/ImageList';
+import ImageGallery from '../components/ImageGallery';
 import { Container } from "@mui/system";
 import { Typography } from "@mui/material";
 import { useTranslation } from 'react-i18next';
@@ -12,7 +12,7 @@ function Gallery() {
     <Container maxWidth='xl'>
       <Typography variant="h1">{t('galleryTitle')}</Typography>
       <Typography variant='body1'>{t('galleryText')}</Typography>
-      <ImageList />
+      <ImageGallery />
     </Container>
   )
 }